Move more common code to base class and add that to be base of other tests
