+++ tests/base.py
@@ -1,21 +1,11 @@
 import random
+import decimal
+import unittest
+import string
+
+from boto3.dynamodb.types import TypeSerializer
 
 class TestDynamoDBBase(unittest.TestCase):
-    def clear_tables(client, table_names):
-        for table_name in table_names:
-            try:
-                table = client.delete_table(TableName=table_name)
-            except Exception as e:
-                print('\nCould not clear table {} due to error {}'.format(table_name, e))
-
-        # wait for all tables to be deleted
-        waiter = client.get_waiter('table_not_exists')
-        for table_name in table_names:
-            try:
-                waiter.wait(TableName=table_name, WaiterConfig={"Delay": 3, "MaxAttempts": 20})        
-            except Exception as e:
-                print('\nCould not wait on table {} due to error {}'.format(table_name, e))
-
     def create_table(client, table_name, hash_key, hash_type, sort_key, sort_type):
         print('\nCreating table: {}'.format(table_name))
 
@@ -44,23 +34,51 @@ class TestDynamoDBBase(unittest.TestCase):
             ProvisionedThroughput={
                 'ReadCapacityUnits': 1,
                 'WriteCapacityUnits': 1
-            }
+            },
+            StreamSpecification={
+                'StreamEnabled': True,
+                'StreamViewType': 'NEW_IMAGE'
+            },
         )
         print('Finished creating table: {}'.format(table_name))
+
+    def clear_tables(client, table_names):
+        for table_name in table_names:
+            try:
+                table = client.delete_table(TableName=table_name)
+            except Exception as e:
+                print('\nCould not clear table {} due to error {}'.format(table_name, e))
+
+        # wait for all tables to be deleted
+        waiter = client.get_waiter('table_not_exists')
+        for table_name in table_names:
+            try:
+                waiter.wait(TableName=table_name, WaiterConfig={"Delay": 3, "MaxAttempts": 20})        
+            except Exception as e:
+                print('\nCould not wait on table {} due to error {}'.format(table_name, e))
 
     def random_string_generator(size=6, chars=string.ascii_uppercase + string.digits):
         return ''.join(random.choice(chars) for x in range(size))
 
-
-    def generate_simple_items_1(num_items):
+    def generate_items(num_items):
+        serializer = TypeSerializer()
         for i in range(num_items):
-            yield {'int_id': { 'N': str(i) },
-                'string_field': {'S': random_string_generator() } }
-
-
-    def generate_simple_items_2(num_items):
-        for i in range(num_items):
-            yield {'string_id': { 'S': random_string_generator() },
-                'int_field': {'N': str(i) } }
+            record = {
+                'int_id': i,
+                'decimal_field': decimal.Decimal(str(i) + '.00000000001'),
+                'string_field': random_string_generator(),
+                'byte_field': b'some_bytes',
+                'int_list_field': [i, i+1, i+2],
+                'int_set_field': set([i, i+1, i+2]),
+                'map_field': {
+                    'map_entry_1': 'map_value_1',
+                    'map_entry_2': 'map_value_2'
+                },
+                'string_list': [random_string_generator(), random_string_generator(), random_string_generator()],
+                'boolean_field': True,
+                'other_boolean_field': False,
+                'null_field': None
+            }
+            yield serializer.serialize(record)
